requirements.txt
kafka-python==2.0.1
psycopg2-binary==2.8
pytest==6.0.1
